renamed 'test_worker.js' to 'testWorker.js' and updated imports
